// src/app/model/products.ts
export class Products {
    ID: number = 0;
    NAME: string = '';
    QUANTITY: number = 0;
    PRICE: number = 0;
    ORI_PRICE: number = 0;
    USD_PRICE: number = 0;
    HOTPRODUCT: number = 0;
    IMAGE: string = '';
    COMPANY: number = 0;
    SEX: number = 0;
    DESCRIPTION: string ='';
    IDCATEGORY: string ='';
  }
  